requirements.txt
pydub==0.25.1
moviepy==1.0.3
Pillow==10.4.0
PyYAML==6.0.2
numpy==1.26.4
srt==3.5.3
requests==2.32.3
